Combined all flight_control nodes into manual_control
