+++ src/surface/flight_control/launch/flight_control_launch.py
@@ -14,30 +14,6 @@ def generate_launch_description() -> LaunchDescription:
         remappings=[
             ('/surface/manipulator_control', '/tether/manipulator_control'),
             ('/surface/mavros/cmd/arming', '/tether/mavros/cmd/arming'),
-        ],
-        emulate_tty=True,
-        output='screen',
-    )
-
-    auto_docking_node = Node(
-        package='flight_control',
-        executable='auto_docking_node',
-        remappings=[('/surface/manipulator_control', '/tether/manipulator_control')],
-        emulate_tty=True,
-        output='screen',
-    )
-
-    control_inverter_node = Node(
-        package='flight_control',
-        executable='control_inverter_node',
-        emulate_tty=True,
-        output='screen',
-    )
-
-    multiplexer_node = Node(
-        package='flight_control',
-        executable='multiplexer_node',
-        remappings=[
             ('/surface/mavros/manual_control/send', '/tether/mavros/manual_control/send'),
             ('/surface/mavros/cmd/command', '/tether/mavros/cmd/command'),
         ],
@@ -46,5 +22,5 @@ def generate_launch_description() -> LaunchDescription:
     )
 
     return LaunchDescription(
-        [manual_control_node, auto_docking_node, control_inverter_node, multiplexer_node]
+        [manual_control_node]
     )
