+++ src/surface/flight_control/setup.py
@@ -28,9 +28,6 @@ setup(
         'console_scripts': [
             'manual_control_node = flight_control.manual_control_node:main',
             'keyboard_control_node = flight_control.keyboard_control_node:main',
-            'auto_docking_node = flight_control.auto_docking_node:main',
-            'control_inverter_node = flight_control.control_inverter_node:main',
-            'multiplexer_node = flight_control.multiplexer:main',
         ],
     },
 )
